+++ src/ewoksorange/gui/concurrency/executor/_EwoksProcessHandle.py
@@ -1,10 +1,12 @@
+from multiprocessing.synchronize import Event
+
 from ._EwoksTaskHandle import EwoksTaskHandle
 
 
 class EwoksProcessHandle(EwoksTaskHandle):
     """Controls a task running in a subprocess via multiprocessing.Event objects."""
 
-    def __init__(self, abort_event, aborted_event):
+    def __init__(self, abort_event: Event, aborted_event: Event):
         self._abort_event = abort_event
         self._aborted_event = aborted_event