+++ src/ewoksorange/gui/concurrency/threaded.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import Optional
 
 from AnyQt.QtCore import QThread
@@ -7,6 +8,14 @@ from .base import TaskExecutor
 
 class ThreadedTaskExecutor(QThread, TaskExecutor):
     """Create and execute an Ewoks task in a dedicated thread."""
+
+    def __init__(self, ewokstaskclass, parent=None):
+        super().__init__(parent=parent, ewokstaskclass=ewokstaskclass)
+        warnings.warn(
+            "ThreadedTaskExecutor is deprecated since version 6.0."
+            "Please use 'ewoksorange.gui.concurrency.executor.EwoksExecutor' instead.",
+            DeprecationWarning,
+        )
 
     def run(self) -> None:
         self.execute_task()