+++ CHANGELOG.md
@@ -7,13 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Fixed 
+### Fixed
 
 - Fix `ParameterForm` file system selection in QT6 (Replace `QDialog.exec_` with `QDialog.exec`).
 
-# Added
+### Added
 
 - `ParameterForm`: add an auto-completer for file / directory selection
+- `ewoksorange.gui.concurrency` add module `executor` for ewoks task execution through thread or sub-processes.
+
+### Changed
+
+- ̀OWEwoksBaseWidget: `execute_ewoks_task` return now when possible a TaskFuture
+- ewoksorange.gui.concurrency.queued.TaskExecutorQueue is deprecated.
+- ewoksorange.gui.concurrency.threaded.ThreadedTaskExecutor is deprecated.
 
 ## [5.0.1] - 2026-04-17